pyproject.toml
[project]
name = "llm-describe-image"
version = "1.0.0"
description = "Task-based image description pipeline using LLM vision models"
readme = "README.md"
requires-python = ">=3.8"
license = {text = "MIT"}
authors = [
    {name = "McDuck"}
]
keywords = ["llm", "image-description", "vision", "ai", "lmstudio"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]

dependencies = [
    "lmstudio>=0.1.0",
    "Pillow>=10.0.0",
    "geopy>=2.0.0",
    "Pillow>=10.0.0",
    "python-dotenv>=1.0.0",
    "tomli>=2.0.0; python_version < '3.11'",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "black>=23.0.0",
    "mypy>=1.0.0",
]

[project.scripts]
llm-describe-image = "llm_describe_directory:main"

[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
py-modules = ["llm_describe_directory", "config_loader"]

[tool.setuptools.packages.find]
where = ["."]
include = ["tasks*"]
